fix(agent): fix postprocessing should drop multi-line completion in fill-in-line mode.
